# openapi/openapi.json
{
  "components": {
    "schemas": {
      "Game": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "description": "Game slug",
            "examples": [
              "space-invaders"
            ],
            "pattern": "^[a-z0-9-]{1,64}$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "Game",
        "type": "object"
      },
      "GameCreate": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "description": "Game slug",
            "examples": [
              "space-invaders"
            ],
            "pattern": "^[a-z0-9-]{1,64}$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "examples": [
              "Space Invaders"
            ],
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "GameCreate",
        "type": "object"
      },
      "GameList": {
        "additionalProperties": false,
        "properties": {
          "games": {
            "items": {
              "$ref": "#/components/schemas/Game"
            },
            "title": "Games",
            "type": "array"
          }
        },
        "required": [
          "games"
        ],
        "title": "GameList",
        "type": "object"
      },
      "HealthStatus": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "HealthStatus",
        "type": "object"
      },
      "Leaderboard": {
        "additionalProperties": false,
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "game_id": {
            "description": "Game slug",
            "examples": [
              "space-invaders"
            ],
            "pattern": "^[a-z0-9-]{1,64}$",
            "title": "Game Id",
            "type": "string"
          },
          "total_players": {
            "minimum": 0.0,
            "title": "Total Players",
            "type": "integer"
          }
        },
        "required": [
          "game_id",
          "total_players",
          "entries"
        ],
        "title": "Leaderboard",
        "type": "object"
      },
      "LeaderboardEntry": {
        "additionalProperties": false,
        "properties": {
          "achieved_at": {
            "format": "date-time",
            "title": "Achieved At",
            "type": "string"
          },
          "rank": {
            "description": "1-based position in the leaderboard",
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "score": {
            "examples": [
              1500
            ],
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "integer"
          },
          "user_id": {
            "description": "User identifier",
            "examples": [
              "player_42"
            ],
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "rank",
          "user_id",
          "score",
          "achieved_at"
        ],
        "title": "LeaderboardEntry",
        "type": "object"
      },
      "LeaderboardEvent": {
        "additionalProperties": false,
        "description": "Payload of each `leaderboard_updated` Server-Sent Event.",
        "properties": {
          "best_score": {
            "examples": [
              1500
            ],
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Best Score",
            "type": "integer"
          },
          "game_id": {
            "description": "Game slug",
            "examples": [
              "space-invaders"
            ],
            "pattern": "^[a-z0-9-]{1,64}$",
            "title": "Game Id",
            "type": "string"
          },
          "user_id": {
            "description": "User identifier",
            "examples": [
              "player_42"
            ],
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "game_id",
          "user_id",
          "best_score"
        ],
        "title": "LeaderboardEvent",
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": true,
        "description": "RFC 9457 problem details (`application/problem+json`).",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Stable machine-readable error code",
            "title": "Code"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "URI identifying the problem type",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProblemError": {
        "additionalProperties": false,
        "description": "A single validation problem. Exactly one of pointer/parameter/header locates it.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "header": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Offending request header",
            "title": "Header"
          },
          "parameter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Offending query or path parameter",
            "title": "Parameter"
          },
          "pointer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "JSON Pointer (RFC 6901) into the request body, e.g. #/score",
            "title": "Pointer"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ProblemError",
        "type": "object"
      },
      "ScoreSubmission": {
        "additionalProperties": false,
        "properties": {
          "score": {
            "examples": [
              1500
            ],
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "integer"
          },
          "user_id": {
            "description": "User identifier",
            "examples": [
              "player_42"
            ],
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "score"
        ],
        "title": "ScoreSubmission",
        "type": "object"
      },
      "ScoreSubmissionResult": {
        "additionalProperties": false,
        "properties": {
          "best_score": {
            "examples": [
              1500
            ],
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Best Score",
            "type": "integer"
          },
          "game_id": {
            "description": "Game slug",
            "examples": [
              "space-invaders"
            ],
            "pattern": "^[a-z0-9-]{1,64}$",
            "title": "Game Id",
            "type": "string"
          },
          "improved": {
            "description": "True if this submission raised the user's best score",
            "title": "Improved",
            "type": "boolean"
          },
          "rank": {
            "description": "1-based position in the leaderboard",
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "submitted_score": {
            "examples": [
              1500
            ],
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Submitted Score",
            "type": "integer"
          },
          "user_id": {
            "description": "User identifier",
            "examples": [
              "player_42"
            ],
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "game_id",
          "user_id",
          "submitted_score",
          "best_score",
          "rank",
          "improved"
        ],
        "title": "ScoreSubmissionResult",
        "type": "object"
      },
      "UserContext": {
        "additionalProperties": false,
        "properties": {
          "above": {
            "description": "Users ranked directly above, best first",
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntry"
            },
            "title": "Above",
            "type": "array"
          },
          "below": {
            "description": "Users ranked directly below, best first",
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntry"
            },
            "title": "Below",
            "type": "array"
          },
          "game_id": {
            "description": "Game slug",
            "examples": [
              "space-invaders"
            ],
            "pattern": "^[a-z0-9-]{1,64}$",
            "title": "Game Id",
            "type": "string"
          },
          "total_players": {
            "minimum": 1.0,
            "title": "Total Players",
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/LeaderboardEntry"
          }
        },
        "required": [
          "game_id",
          "total_players",
          "user",
          "above",
          "below"
        ],
        "title": "UserContext",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Real-time gaming leaderboard API. Errors use RFC 9457 problem details.",
    "title": "lbserv",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/health": {
      "get": {
        "description": "Liveness probe (no dependencies).",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/api/ready": {
      "get": {
        "description": "Readiness probe: verifies database connectivity.",
        "operationId": "getReadiness",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Ready",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/games": {
      "get": {
        "operationId": "listGames",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameList"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Games",
        "tags": [
          "games"
        ]
      },
      "post": {
        "operationId": "createGame",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Create Game",
        "tags": [
          "games"
        ]
      }
    },
    "/api/v1/games/{game_id}/events": {
      "get": {
        "operationId": "streamLeaderboardEvents",
        "parameters": [
          {
            "description": "Game slug",
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "description": "Game slug",
              "pattern": "^[a-z0-9-]{1,64}$",
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardEvent"
                }
              }
            },
            "description": "Server-Sent Events stream; each `leaderboard_updated` event's data is a LeaderboardEvent JSON document."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Stream Leaderboard Events",
        "tags": [
          "leaderboard"
        ]
      }
    },
    "/api/v1/games/{game_id}/leaderboard": {
      "get": {
        "description": "Top users for a game, best first.",
        "operationId": "getLeaderboard",
        "parameters": [
          {
            "description": "Game slug",
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "description": "Game slug",
              "pattern": "^[a-z0-9-]{1,64}$",
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "description": "Number of top users",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "description": "Number of top users",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Leaderboard"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Leaderboard",
        "tags": [
          "leaderboard"
        ]
      }
    },
    "/api/v1/games/{game_id}/scores": {
      "post": {
        "description": "Submit a score. The leaderboard keeps each user's best score per game.",
        "operationId": "submitScore",
        "parameters": [
          {
            "description": "Game slug",
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "description": "Game slug",
              "pattern": "^[a-z0-9-]{1,64}$",
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreSubmissionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Submit Score",
        "tags": [
          "leaderboard"
        ]
      }
    },
    "/api/v1/games/{game_id}/users/{user_id}/context": {
      "get": {
        "description": "A user's current rank with the users immediately above and below them.",
        "operationId": "getUserContext",
        "parameters": [
          {
            "description": "Game slug",
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "description": "Game slug",
              "pattern": "^[a-z0-9-]{1,64}$",
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "description": "User identifier",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "description": "User identifier",
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "description": "Users to include above and below",
            "in": "query",
            "name": "neighbors",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Users to include above and below",
              "maximum": 10,
              "minimum": 0,
              "title": "Neighbors",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserContext"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get User Context",
        "tags": [
          "leaderboard"
        ]
      }
    }
  }
}